refactor, add ubuntu14 mount util
